Restrucured map components
- Made the structure map responsible for drawing/updating its objects. This is quite handy, since we can now call StructureMap.AddPlatform and it handles adding it to the fog of war and stuff.

- Made the camera "independet" from the map. The only reason we had it in the map was to pass an update, which is not needed anymore since the input manager passes all the input stuff to the camera.

- Minor changes to the levels, to pass the new map structure correctly

+++ Singularity/Singularity/Platform/PlatformFactory.cs
@@ -11,6 +11,9 @@ using Singularity.Map;
 
 namespace Singularity.Platform
 {
+    /// <summary>
+    /// The basic idea for this was to provide an easy way to get a platform via the platformtype.
+    /// </summary>
     static class PlatformFactory
     {
         private static Texture2D mConeSheet;
@@ -23,7 +26,13 @@ namespace Singularity.Platform
 
         private static Director mDirector;
 
-
+        /// <summary>
+        /// Initializes the platform factory with the sprite sheets needed.
+        /// </summary>
+        /// <param name="coneSheet">The cone sprite sheet for the facilities</param>
+        /// <param name="cylinderSheet">The cylinder sprite sheet for the facilities</param>
+        /// <param name="domeSheet">The dome sprite sheet for the facilities</param>
+        /// <param name="blankSheet">The blank sprite sheet for the facilities</param>
         public static void Init(Texture2D coneSheet, Texture2D cylinderSheet, Texture2D domeSheet, Texture2D blankSheet)
         {
             mConeSheet = coneSheet;
@@ -32,7 +41,16 @@ namespace Singularity.Platform
             mBlankSheet = blankSheet;
         }
 
-        public static PlatformBlank Get(EPlatformType type, ref Director director, int x = 0, int y = 0, ResourceMap resourceMap = null)
+        /// <summary>
+        /// Returns the platform specified by the given platform type.
+        /// </summary>
+        /// <param name="type">The type of the platform to get</param>
+        /// <param name="director">A reference to the director needed to initializes some platforms</param>
+        /// <param name="x">The initial x coordinate of the platform</param>
+        /// <param name="y">The initial y coordinate of the platform</param>
+        /// <param name="resourceMap">The resource map needed to initialize some platforms</param>
+        /// <returns></returns>
+        public static PlatformBlank Get(EPlatformType type, ref Director director, float x = 0, float y = 0, ResourceMap resourceMap = null)
         {
             //TODO: add conesheet to this query. Its not included right now since it doesn't exists at this point in time
             if (mCylinderSheet == null || mDomeSheet == null || mBlankSheet == null)
